requirements.txt
pyyaml==6.0.3
python-docx==1.1.2
langgraph==0.2.40
pydantic>=2.9.0,<3
requests==2.32.5
google-genai==1.66.0
fastapi==0.115.6
uvicorn==0.34.0
python-dotenv==1.0.1
python-multipart==0.0.20
PyPDF2==3.0.1
groq==0.15.0